+++ src/amr/solvers/solver_field_evolvers.hpp
@@ -130,6 +130,28 @@ OhmLevelTransformer(core::OhmInfo, typename Model::amr_types::level_t&, Model&)
     -> OhmLevelTransformer<Model>;
 
 
+
+
+
+template<typename level_t, typename Model>
+struct TimeSetter
+{
+    void operator()(auto&... quantities)
+    {
+        auto& rm = *model.resourcesManager;
+        for (auto& patch : rm.enumerate(level, quantities...))
+            (model.resourcesManager->setTime(quantities, *patch, newTime), ...);
+    }
+
+    level_t& level;
+    Model& model;
+    double newTime;
+};
+
+template<typename level_t, typename Model>
+TimeSetter(level_t&, Model&, double) -> TimeSetter<level_t, Model>;
+
+
 } // namespace PHARE::solver
